indrz/requirements/base.txt
# Django
# ------------------------------------------------------------------------------
django==4.2.16
django-mptt==0.14.0
django-taggit==3.1.0
django-cors-headers==3.13.0
djangorestframework_gis==1.0
djangorestframework==3.15.2
python3-saml-django==1.3.0
drf-yasg==1.21.7
tzdata==2024.1

# Data
# ------------------------------------------------------------------------------
pandas==1.5.2
geojson==2.5.0
requests==2.32.3
pillow==10.3.0
psycopg2-binary==2.9.9

